requirements.txt
Flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
ta==0.11.0
joblib==1.4.2